requirements.txt
yt-dlp>=2023.0.0
openai>=1.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
bilibili-api-python>=17.0.0
requests>=2.28.0
httpx>=0.24.0
lark-oapi>=1.0.0
